# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pelayanan Scheduler API",
    "description": "API for managing church service scheduling including users, positions, categories, and worship services",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": ""
      }
    },
    "schemas": {
...
      "CreateIbadah": {
        "type": "object",
        "required": ["id_users", "id_ibadahCategory", "id_pelayanPosition"],
        "properties": {
          "id_ibadahCategory": { "type": "integer", "minimum": 1, "example": 1 },
          "stola": { "type": "string", "example": "ungu" },
          "dress_code": { "type": "string", "example": "batik" },
          "service_date": { "type": "string", "format": "date" },
          "tata_ibadah_link": {"type": "string", "example": "https://drive.google.com/file/d/1KfLa8w9lm34fTWi_l9ZtxOkLz6eZG8Qy/view?usp=sharing"},
          "start_service_time": { "type": "string", "format": "date-time", "example": "08:00" },
          "end_service_time": { "type": "string", "format": "date-time", "example": "09:00" }
        }
      },
      "UpdateIbadah": {
        "type": "object",
        "properties": {
          "id_users": { "type": "integer", "minimum": 1, "example": 1 },
          "id_ibadahCategory": { "type": "integer", "minimum": 1, "example": 1 },
          "id_pelayanPosition": { "type": "integer", "minimum": 1, "example": 1 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Error message" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": { "type": "string" },
                "expected": { "type": "string" },
                "received": { "type": "string" },
                "path": { "type": "array", "items": { "type": "string" } },
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/ibadah": {
      "get": {
        "summary": "Get all ibadah schedules",
        "description": "Retrieve all ibadah with their related data",
        "tags": ["Ibadahs"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "required": false,
            "description": "Search ibadah by name"
          },
          {
            "in": "query",
            "name": "id_ibadahCategory",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by Ibadah Category"
          },
          {
            "in": "query",
            "name": "id_pelayanPosition",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by Pelayan Position ID"
          },
          {
            "in": "query",
            "name": "id_pelayanLevel",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by pelayan level ID"
          },
          {
            "in": "query",
            "name": "id_krw",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by KRW ID"
          },
          {
            "in": "query",
            "name": "skip",
            "schema": { "type": "integer", "default": 0 },
            "required": false,
            "description": "Number of records to skip (for pagination)"
          },
          {
            "in": "query",
            "name": "take",
            "schema": { "type": "integer", "default": 10, "maximum": 100 },
            "required": false,
            "description": "Number of records to return (max 100)"
          }
         ],
        "responses": {
          "200": {
            "description": "List of ibadah schedules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Ibadah" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new ibadah schedule",
        "description": "Create a new user with required relationships",
        "tags": ["Ibadahs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateIbadah" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ibadah schedule created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Ibadah" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or referenced record not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/ibadah/{id}": {
      "get": {
        "summary": "Get ibadah schedule by ID",
        "description": "Retrieve a specific user by their ID with related data",
        "tags": ["Ibadahs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "Ibadah ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ibadah schedule data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Ibadah" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid user ID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a ibadah schedule",
        "description": "Update ibadah  information",
        "tags": ["Ibadahs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateIbadah" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ibadah updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Ibadah" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid references",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a ibadah schedule",
        "description": "Delete a ibadah schedule by ID",
        "tags": ["Ibadahs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "Ibadah ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ibadah deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Ibadah deleted successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ibadah ID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Ibadah not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/ibadah-assignments": {
      "get": {
        "summary": "Get all ibadah assignments",
        "description": "Retrieve all ibadah assignments with their related data (user, position, ibadah schedule)",
        "tags": ["IbadahAssignments"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "required": false,
            "description": "Search assignments by user name, position, or ibadah category"
          },
          {
            "in": "query",
            "name": "id_ibadah",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by Ibadah ID"
          },
          {
            "in": "query",
            "name": "id_pelayanPosition",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by Pelayan Position ID"
          },
          {
            "in": "query",
            "name": "id_users",
            "schema": { "type": "integer" },
            "required": false,
            "description": "Filter by User ID"
          },
          {
            "in": "query",
            "name": "skip",
            "schema": { "type": "integer", "default": 0 },
            "required": false,
            "description": "Number of records to skip (for pagination)"
          },
          {
            "in": "query",
            "name": "take",
            "schema": { "type": "integer", "default": 10, "maximum": 100 },
            "required": false,
            "description": "Number of records to return (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of ibadah assignments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/IbadahAssignment" }
                    },
                    "total": { "type": "integer", "example": 42 },
                    "skip": { "type": "integer", "example": 0 },
                    "take": { "type": "integer", "example": 10 }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new ibadah assignment",
        "description": "Assign a user to an ibadah schedule with a position",
        "tags": ["IbadahAssignments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateIbadahAssignment" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ibadah assignment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/IbadahAssignment" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or referenced record not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}